backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pymongo
requests
pillow
gTTS
groq
python-multipart
torch
transformers
fpdf
passlib==1.7.4 
bcrypt==3.2.2
pinecone
sentence-transformers
numpy
pypdf
open_clip_torch
